requirement.txt
numpy==1.19.4
Pillow==8.0.1
PyYAML==5.3.1
six==1.15.0
wxPython==4.1.0
